Added JS Wrapper for FML Event
This class sholud wrap any FML Event methods that would want to be
accessed from JavaScript with more JS Friendly versions. This should
also be used to ensure backward compatability over time.

the original Event is stored in the public variable _event so that it
conforms to the JS standard for variables for internal use only. but is
accessable in case a more advanced programmer wants to access the event
for some reason.